% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{boot_smoothing_error}
\alias{boot_smoothing_error}
\title{Calculate bootstrapped standard smoothing error}
\usage{
boot_smoothing_error(time_period_result, data)
}
\arguments{
\item{time_period_result}{output from  \code{forecast_time_period}}

\item{data}{\emph{data frame} at each time-step containing two columns: date and confirm (number of cases per day)}
}
\value{
bootstrapped smoothing error for each sliding window
}
\description{
Calculate bootstrapped standard smoothing error
}
