% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{check_match_names}
\alias{check_match_names}
\title{Check Matching Names in Named Vectors}
\usage{
check_match_names(...)
}
\arguments{
\item{...}{Named numeric vectors to be compared.}
}
\value{
Throws an error if name mismatches are found; otherwise, returns nothing.
}
\description{
Internal utility function to verify that all input named vectors share the same names.
}
\keyword{internal}
