% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_df_pred_case}
\alias{combine_df_pred_case}
\title{Extract combined dataframe with forecast quantiles and weekly case data for summary function}
\usage{
combine_df_pred_case(forecast_dat, data, pred_horizon_str = NULL)
}
\arguments{
\item{forecast_dat}{output from  \code{create_forecast_df}}

\item{data}{\emph{data frame} containing two columns: date and confirm (number of cases per week)}

\item{pred_horizon_str}{\emph{string} prediction horizon time period to plot}
}
\value{
combined dataframe with forecast quantiles and weekly case data for summary function
#' \describe{
\item{date}{daily date or weekly date over which samples were aggregated}
\item{p50}{median quantile value}
\item{p25}{quantile value of probability 0.25}
\item{p75}{quantile value of probability 0.75}
\item{p05}{quantile value of probability 0.05}
\item{p95}{quantile value of probability 0.95}
\item{pred_horizon}{Prediction time horizon (time period ahead for which prediction was made)}
\item{sim_draws}{prediction values from simulated draws}
\item{min_sim}{minimum value of simulated predictions for week}
\item{max_sim}{maximum value of simulated predictions for week}
\item{confirm}{confirmed weekly cases}
}
}
\description{
Extract combined dataframe with forecast quantiles and weekly case data for summary function
}
