% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{common_reliable_estimation_date}
\alias{common_reliable_estimation_date}
\title{Extract common date that all diseases in PHRDW and PLOVER data can be reliably estimated with in EpiEstim}
\usage{
common_reliable_estimation_date(
  plover_list,
  phrdw_list,
  flua_min = config$min_nb_cases_flua,
  flub_min = config$min_nb_cases_flub,
  cov_min = config$min_nb_cases_covid,
  rsv_min = config$min_nb_cases_rsv
)
}
\arguments{
\item{plover_list}{List of transformed PLOVER data (output of \code{get_all_vri_data})}

\item{phrdw_list}{List of transformed PHRDW data (output of \code{get_all_vri_data})}

\item{flua_min}{Minimum number of Flu-A cases required for reliable estimation by EpiEstim}

\item{flub_min}{Minimum number of Flu-B cases required for reliable estimation by EpiEstim}

\item{cov_min}{Minimum number of SARS-CoV2 cases required for reliable estimation by EpiEstim}

\item{rsv_min}{Minimum number of RSV cases required for reliable estimation by EpiEstim}
}
\value{
\emph{numeric} Minimum reliable common date
}
\description{
Extract common date that all diseases in PHRDW and PLOVER data can be reliably estimated with in EpiEstim
}
