% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_epiestim_model_functions.R
\name{create_1w_2w_df_epiestim}
\alias{create_1w_2w_df_epiestim}
\title{Extract dataframe with 1 week and 2 week ahead forecasts for Violin plot}
\usage{
create_1w_2w_df_epiestim(time_period_result)
}
\arguments{
\item{time_period_result}{output from  \code{forecast_time_period_epiestim}}
}
\value{
Data frame with 1 week and 2 week forecasts at all time-points
}
\description{
Extract dataframe with 1 week and 2 week ahead forecasts for Violin plot
}
