% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_epiestim_model_functions.R
\name{create_forecast_df}
\alias{create_forecast_df}
\title{Extract data-frame with forecasts for violin plot}
\usage{
create_forecast_df(time_period_result)
}
\arguments{
\item{time_period_result}{output from  \code{forecast_time_period_epiestim}}
}
\value{
Data frame with forecasts at all time-points
}
\description{
Extract data-frame with forecasts for violin plot
}
