% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_quantiles}
\alias{create_quantiles}
\title{summarise a data frame \code{d} by groups along a \code{variable}}
\usage{
create_quantiles(d, ..., variable = NULL)
}
\arguments{
\item{d}{tibble data frame}

\item{...}{group_by variables}

\item{variable}{string

@importFrom magrittr "\%>\%"
@import dplyr}
}
\description{
summarise a data frame \code{d} by groups along a \code{variable}
}
