% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daily_data}
\alias{daily_data}
\title{Daily/Weekly flu, RSV, sars_cov2 Cases}
\format{
A data frame with over 366 rows and 4 variables:
\describe{
\item{date}{YMD format}
\item{rsv}{number of confirmed RSV cases}
\item{flua}{number of confirmed flu A cases}
\item{covid}{number of confirmed COVID-19 cases}
}
}
\usage{
daily_data
}
\description{
A simulated dataset containing COVID, flu A and RSV cases
The variables are as follows:
}
\keyword{datasets}
