% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extend_rows_model_data}
\alias{extend_rows_model_data}
\title{extract model data with extension during each iteration of loop}
\usage{
extend_rows_model_data(data, min_model_date_str, extension_interval = 0)
}
\arguments{
\item{min_model_date_str}{start date (in str)}

\item{extension_interval}{an integer (# of days)

#'  @importFrom magrittr "\%>\%"
@import dplyr
@import lubridate}
}
\value{
model_data with # of days extension specified
}
\description{
extract model data with extension during each iteration of loop
}
