% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_agg_samples_epiestim_fit}
\alias{extract_agg_samples_epiestim_fit}
\title{Summarise incidence values into weekly aggregate}
\usage{
extract_agg_samples_epiestim_fit(samples)
}
\arguments{
\item{samples}{Daily samples generated from \code{extract_daily_samples_epiestim_fit}}
}
\description{
Summarise incidence values into weekly aggregate
}
