% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_epiestim_model.R
\name{extract_daily_samples_epiestim_fit}
\alias{extract_daily_samples_epiestim_fit}
\title{Extract daily forecast samples}
\usage{
extract_daily_samples_epiestim_fit(
  data,
  model_fit,
  dt = 7L,
  n_days = 14,
  n_sim = 1000,
  ...
)
}
\arguments{
\item{data}{\emph{data frame} containing two columns: date and confirm (number of cases per day)}

\item{model_fit}{Object of class {\code{\link[EpiEstim]{estimate_R}}} generated by running \code{fit_epiestim_model}
@param dt \emph{Integer} Number of days aggregated (set to 7 by default for weekly aggregate)
@param n_days     The number of days to run simulations for. Defaults to 14
@param n_sim The number of epicurves to simulate. Defaults to 1000
@param ... Optional parameters to pass on to {\code{\link[projections]{project}}} (see help page)}
}
\value{
Data-frame of daily forecast samples from all simulations
}
\description{
Function to produce short-term daily projections from objects of class {\code{\link[EpiEstim]{estimate_R}}}
}
