% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_epiestim_model.R
\name{extract_experiment_model_data_epiestim}
\alias{extract_experiment_model_data_epiestim}
\title{Extract the last experiment model data (aka all the fitted data in this experiment)}
\usage{
extract_experiment_model_data_epiestim(time_period_result)
}
\arguments{
\item{time_period_result}{output from experiment_time_period_epiestim()}
}
\description{
Extract the last experiment model data (aka all the fitted data in this experiment)
}
