% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_quantile_epiestim}
\alias{extract_quantile_epiestim}
\title{Extract calculated quantiles from the weekly samples}
\usage{
extract_quantile_epiestim(tp)
}
\arguments{
\item{tp}{element from list output generated by \code{forecast_time_period}}
}
\description{
Extract calculated quantiles from the weekly samples
}
