% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_sim_samples_epiestim}
\alias{extract_sim_samples_epiestim}
\title{Extract simulated samples ideally 1000 samples per date}
\usage{
extract_sim_samples_epiestim(tp, aggregate_unit = NULL)
}
\arguments{
\item{tp}{element from list output generated by \code{forecast_time_period_epiestim}}

\item{aggregate_unit}{Time forecasted samples are aggregated by (weekly or daily)}
}
\description{
Extract simulated samples ideally 1000 samples per date
}
