% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_weekly_quantile_epiestim}
\alias{extract_weekly_quantile_epiestim}
\title{Extract calculated quantiles from the weekly samples}
\usage{
extract_weekly_quantile_epiestim(time_period_result)
}
\arguments{
\item{time_period_result}{output from \code{forecast_time_period_epiestim}}
}
\description{
Extract calculated quantiles from the weekly samples
}
