% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_weekly_samples_epiestim}
\alias{extract_weekly_samples_epiestim}
\title{Extract Weekly Samples (week_date & value) ideally 1000 samples per date}
\usage{
extract_weekly_samples_epiestim(time_period_result)
}
\arguments{
\item{time_period_result}{output from \code{forecast_time_period_epiestim}}
}
\description{
Extract Weekly Samples (week_date & value) ideally 1000 samples per date
}
