% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_dates}
\alias{filter_dates}
\title{Helper function to filter and extract first reliable date}
\usage{
filter_dates(data, min_cases)
}
\arguments{
\item{data}{weekly transformed PLOVER or PHRDW data}

\item{min_cases}{minimum number of reliable cases for EpiEstim from config file}
}
\description{
Helper function to filter and extract first reliable date
}
