% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_epiestim_model_functions.R
\name{fit_epiestim_model}
\alias{fit_epiestim_model}
\title{fit_epiestim_model - Function to estimate the reproduction number of an epidemic from weekly data}
\usage{
fit_epiestim_model(
  data,
  dt = 7L,
  type = NULL,
  mean_si = NULL,
  std_si = NULL,
  recon_opt = "match",
  method = "parametric_si",
  ...
)
}
\arguments{
\item{data}{\emph{data frame} containing two columns: date and confirm (number of cases per week)}

\item{dt}{\emph{Integer}     length of temporal aggregations of the incidence data. This should be an integer or vector of integers. The default value is 7 time units (1 week).}

\item{type}{\emph{character} Specifies type of epidemic. Must be one of "flu_a", "flu_b", "rsv", "covid" or "other"}

\item{mean_si}{\emph{Numeric} User specification of mean of parametric serial interval}

\item{std_si}{\emph{Numeric} User specification of standard deviation of parametric serial interval}

\item{recon_opt}{One of "naive" or "match" to pass on to {\code{\link[EpiEstim]{estimate_R}}} (see help page)}

\item{method}{One of "non_parametric_si", "parametric_si", "uncertain_si", "si_from_data" or "si_from_sample" to pass on to {\code{\link[EpiEstim]{estimate_R}}} (see help page)}

\item{...}{Other optional parameters to pass on to {\code{\link[EpiEstim]{estimate_R}}} (see help page) to control estimation of reproduction number}
}
\value{
Object of class {\code{\link[EpiEstim]{estimate_R}}} (see \code{EpiEstim} help page)
}
\description{
A wrapper function for {\code{\link[EpiEstim]{estimate_R}}} from the \code{EpiEstim} library to estimate the reproduction number of epidemics to support short-term forecasts
}
\details{
\code{fit_epiestim_model} currently supports the following epidemics: Influenza, RSV and COVID-19. The serial intervals for the estimation of R were retrieved from
Cowling et al., 2011, Vink et al., 2014 and Madewell et al., 2023 for Influenza, RSV and COVID (BA.5 Omicron variant) respectively
}
\examples{
plover_data <- data.frame(
          epiWeek_date = as.Date(c('2022-10-02', '2022-10-09',
                     '2022-10-16', '2022-10-23', '2022-10-30',
                     '2022-11-06', '2022-11-13', '2022-11-20',
                      '2022-11-27', '2022-12-04')),
          epiWeek_year = c(2022,2022,2022,2022,
                      2022,2022,2022,2022,2022,2022),
           Epiweek = c(40,41,42,43,44,45,46,47,48,49),
           flu_a = c(17,19,32,38,43,45,73,88,94,105),
           flu_b = c(24,31,39,45,50,52,68,83,89,97))

weekly_plover_data <- get_weekly_plover(plover_data)

plover_dat_clean <- get_weekly_plover_by_date_type(
                             weekly_plover_data,
                             'flu_a',
                             '2022-10-01',
                             '2022-12-05')

fit_epiestim_model(data = plover_dat_clean, type = "flu_a")
}
