% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_epifilter_model_functions.R
\name{fit_epifilter_model}
\alias{fit_epifilter_model}
\title{fit_epifilter_model - Function to estimate the reproduction number of an epidemic}
\usage{
fit_epifilter_model(
  data,
  dt = 1L,
  type = NULL,
  mean_si = NULL,
  std_si = NULL,
  grid_size = 200,
  Rmin = 0.01,
  Rmax = 3,
  smooth = TRUE,
  confidence_est = 0.025,
  eta = 0.1
)
}
\arguments{
\item{data}{\emph{data frame} containing two columns: date and confirm (number of cases per week)}

\item{dt}{\emph{Integer}     length of temporal aggregations of the incidence data. This should be an integer or vector of integers. The default value for epiFilter is 1L (daily data)}

\item{type}{\emph{character} Specifies type of epidemic. Must be one of "flu_a", "flu_b", "rsv", "sars_cov2" or "other"}

\item{mean_si}{\emph{numeric} User specification of mean of parametric serial interval}

\item{std_si}{\emph{numeric} User specification of standard deviation of parametric serial interval}

\item{grid_size}{\emph{numeric} Size of uniform grid for reproduction number values (set to 200 by default)}

\item{smooth}{\emph{logical} set to TRUE by default to apply recursive Bayesian smoothing of estimates}

\item{confidence_est}{\emph{numeric} Confidence level of Rhat estimates (set to 0.025 yielding 95\% confidence intervals by default)}

\item{eta}{\emph{numeric} diffusion noise (set to 0.1 by default)}

\item{mean_prior}{\emph{numeric} positive number for the minimum grid reproduction number}

\item{std_prior}{\emph{numeric} positive number for the maximum grid reproduction number}
}
\value{
List storing filtered Rhat estimates and model outputs
\describe{
\item{Rmedian}{50\% quantiles of Rhat estimates}
\item{Rhat_estimates}{95\% quantiles of Rhat estimates}
\item{Rmean}{Average of Rhat estimates}
\item {smooth_filtered_posterior}{Smoothed and filtered posterior estimates}
}
\description{
Model fitting function using recursive Bayesian filtering to estimate the reproduction number of epidemics to support short-term forecasts
From: Parag, KV, (2020) “Improved real-time estimation of reproduction numbers
}
\details{
\code{fit_epifilter_model} currently supports the following epidemics: Influenza, RSV and SARS-CoV2. The serial intervals for the estimation of R were retrieved from
Cowling et al., 2011, Vink et al., 2014 and Madewell et al., 2023 for Influenza A, Influenza B, RSV and SARS-CoV2 (BA.5 Omicron variant) respectively
}
\examples{
fit_epifilter_model(data = plover_data, type = "flu_a")

}
