% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_epiestim_model_functions.R
\name{forecast_epiestim}
\alias{forecast_epiestim}
\title{Main function to run the forecast model}
\usage{
forecast_epiestim(
  data,
  start_date,
  window_size = 7,
  n_days = 7,
  type = NULL,
  smooth_data = FALSE,
  smoothing_cutoff = 10,
  ...
)
}
\description{
Main function to run the forecast model
}
