% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{forecast_metrics}
\alias{forecast_metrics}
\title{Extract current forecast metrics: forecast prediction, percentile interval and Rt value}
\usage{
forecast_metrics(time_period_result, iter = 10)
}
\arguments{
\item{time_period_result}{output from  \code{forecast_time_period}}

\item{iter}{number of MCMC iterations used to generate Rt posterior}
}
\value{
current forecast metrics
}
\description{
Extract current forecast metrics: forecast prediction, percentile interval and Rt value
}
