% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{forecast_quality_precheck}
\alias{forecast_quality_precheck}
\title{Get proportion of cases below minimum in confirmed case data numbers as deterministic check for forecast quality}
\usage{
forecast_quality_precheck(data, model_data)
}
\arguments{
\item{data}{\emph{data frame} at each time-step containing two columns: date and confirm (number of cases per day)}

\item{model_data}{input dataframe containing two columns: date and confirm (number of cases per day)}
}
\value{
numeric proportion of zeroes
}
\description{
Get proportion of cases below minimum in confirmed case data numbers as deterministic check for forecast quality
}
