% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_functions.R
\name{forecast_time_period}
\alias{forecast_time_period}
\title{Iterate through a time-period as a sliding window to produce short-term forecasts}
\usage{
forecast_time_period(
  data,
  start_date,
  n_days = 7,
  time_period = "weekly",
  type = NULL,
  algorithm = "EpiEstim"
)
}
\arguments{
\item{data}{\emph{data frame} containing two columns: date and confirm (number of cases per week)}

\item{start_date}{Initial starting time-point. Must match a timepoint in the input dataset}

\item{n_days}{The number of days to run simulations for. Defaults to 7}

\item{time_period}{time period string (e.g. 'daily', 'weekly'). Default is daily}

\item{type}{\emph{character} Specifies type of epidemic. Must be one of "flu_a", "flu_b", "rsv", "sars_cov2" or "other"}

\item{algorithm}{argument to control if model fitting and forecasting is done by \code{EpiEstim} or \code{EpiFilter}}
}
\value{
List of class \code{forecast_time_period}
storing quantiles of both daily and weekly forecasts from each sliding window
}
\description{
Function to produce short-term forecasts using either the EpiEstim or EpiFilter algorithm
}
\examples{

#  Forecast using EpiEstim
forecast_time_period(
  data = weekly_transformed_plover_data,
  start_date = "2022-10-02", n_days = 14, type = "flu_a", algorithm = "EpiEstim"
)

#  Forecast using EpiFilter
forecast_time_period(
  data = weekly_transformed_plover_data,
  start_date = "2022-10-02", n_days = 14, type = "flu_a", algorithm = "EpiFilter"
)
}
