% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_epiestim_model.R
\name{forecast_time_period_epiestim}
\alias{forecast_time_period_epiestim}
\title{Iterate through a time-period as a sliding window to forecast 1-2 weeks ahead with the EpiEstim model fit}
\usage{
forecast_time_period_epiestim(
  data,
  start_date_str,
  time_period = c(0, 7, 14, 21),
  type = NULL
)
}
\arguments{
\item{data}{\emph{data frame} containing two columns: date and confirm (number of cases per day)}

\item{start_date_str}{Initial starting time-point. Must match a timepoint in the input dataset
@param time_period \emph{vector} of multiples of 7 (for weekly forecast data) with the number of weeks forecasts are to be generated for. Default is c(0, 7, 14, 21) indicating that last forecasting will be done from a time-point 3 weeks ahead of the start date specified
@param type \emph{character} Specifies type of epidemic. Must be one of "Influenza", "RSV" and "COVID"}
}
\value{
List storing quantiles of 2 week ahead weekly forecasts from each sliding window
}
\description{
Function to produce short-term weekly forecasts from objects of class {\code{\link[EpiEstim]{estimate_R}}}
}
