% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gam_prediction_error}
\alias{gam_prediction_error}
\title{Calculate GAM prediction error}
\usage{
gam_prediction_error(time_period_result)
}
\arguments{
\item{time_period_result}{output from  \code{forecast_time_period}}
}
\description{
Calculate GAM prediction error
}
