% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{generic_data}
\alias{generic_data}
\title{Daily Generic Confirmed Cases Data}
\format{
\describe{
\item{date_of_report}{the date of the reported cases (e.g. as.Date('2022-01-01'))}
\item{flu_a}{number of confirmed flu A cases}
\item{flu_b}{number of confirmed flu B cases}
}
}
\usage{
generic_data
}
\description{
A dataset containing a generic naming for the date & confirmed cases column
The variables are as follows:
}
\keyword{datasets}
