% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_all_vri_data}
\alias{get_all_vri_data}
\title{Process PLOVER and PHRDW data for all respiratory diseases supported by ViroReportR}
\usage{
get_all_vri_data(raw_plover_data, raw_phrdw_data)
}
\arguments{
\item{raw_plover_data}{PLOVER data}

\item{raw_phrdw_data}{PHRDW data}
}
\value{
Lists of dataframes of transformed PLOVER and PHRDW data for each respiratory viral disease with columns \code{date} and \code{confirm}
}
\description{
Process PLOVER and PHRDW data for all respiratory diseases supported by ViroReportR
}
