% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation_phrdw.R
\name{get_phrdw_by_type_date_age}
\alias{get_phrdw_by_type_date_age}
\title{Filter phrdw data filtered by time period, disease type, date, age}
\usage{
get_phrdw_by_type_date_age(
  phrdw_data,
  time_period = "daily",
  type,
  start_date,
  end_date,
  start_age = 0,
  end_age = 150
)
}
\arguments{
\item{phrdw_data}{raw phrdw data}

\item{time_period}{time period string (e.g. 'daily', 'weekly')}

\item{type}{disease type (e.g. 'sars_cov2', 'rsv', 'flu_a', 'flu_b')}

\item{start_date}{start date string (e.g. '2022-01-01')}

\item{end_date}{end date string (e.g. '2022-12-31')}

\item{start_age}{start age integer (default = 0)}

\item{end_age}{end age integer (default = 150)}
}
\value{
phrdw data filtered by time period, disease type, date, age
}
\description{
\code{get_phrdw_by_type_date_age()} filter the phrdw data by a given
time_period, disease type, date range, and age range, it performs the following steps:
\enumerate{
\item Extracts the phrdw data by daily or weekly
\item Filters the phrdw data by the given epiweek date range and age range.
\item Aggregates the confirmed cases count by the epiweek date.
\item Selects the date & the corresponding disease type column.
\item Renames the disease type as \code{confirm}.
}

The input dataframe \code{phrdw_data} must have the following columns:
\itemize{
\item \code{lis_date_collection}: Date of collection (e.g. '2019-01-01')
\item \code{result_lab_name}: The name of the result lab (e.g. 'BCCDC')
\item \code{age_years}: Age (e.g. 1, 2, ...)
\item \code{sars_cov2}: Confirmed Cases Count (e.g. 1, 2, ...)
\item \code{rsv}: Confirmed Cases Count (e.g. 1, 2, ...)
\item \code{flu_a}: Confirmed Cases Count (e.g. 1, 2, ...)
\item \code{flu_b}: Confirmed Cases Count (e.g. 1, 2, ...)
}

The input disease type must be either 'sars_cov2', 'rsv', 'flu_a', or 'flu_b'.

The input start date must be earlier than the input end date.

The input start age must be earlier than the input end age.
}
\examples{
phrdw_data <- data.frame(
  lis_date_collection = as.Date(c(
    "2022-10-02", "2022-10-02",
    "2022-10-09", "2022-10-09",
    "2022-10-16", "2022-10-16",
    "2022-10-23", "2022-10-23",
    "2022-10-30", "2022-10-30"
  )),
  result_lab_name = c(
    "BC Centre for Disease Control", "BC Centre for Disease Control",
    "BC Centre for Disease Control", "BC Centre for Disease Control",
    "BC Centre for Disease Control", "BC Centre for Disease Control",
    "BC Centre for Disease Control", "BC Centre for Disease Control",
    "BC Centre for Disease Control", "BC Centre for Disease Control"
  ),
  age_years = c(7, 8, 7, 8, 7, 8, 7, 8, 7, 8),
  sars_cov2 = c(15, 17, 30, 28, 41, 43, 71, 74, 80, 88),
  rsv = c(14, 16, 29, 27, 40, 42, 70, 73, 79, 87),
  flu_a = c(13, 15, 28, 26, 39, 41, 69, 72, 78, 86),
  flu_b = c(12, 14, 27, 25, 38, 40, 68, 71, 77, 85)
)


get_phrdw_by_type_date_age(
  phrdw_data = phrdw_data,
  time_period = "weekly",
  type = "rsv",
  start_date = "2022-10-01",
  end_date = "2022-11-01",
  start_age = 0,
  end_age = 10
)
}
