% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation_generic.R
\name{get_weekly_aggregated_data}
\alias{get_weekly_aggregated_data}
\title{Extract Aggregated Weekly Generic Data}
\usage{
get_weekly_aggregated_data(
  generic_data,
  date_column,
  number_column,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{generic_data}{the weekly generic data from \code{get_weekly_data()}}

\item{date_column}{date column name str}

\item{number_column}{cases count column name str}

\item{start_date}{start date string (e.g. '2022-01-01')(optional, default is NULL)}

\item{end_date}{end date string (e.g. '2022-12-31')(optional, default is NULL)}
}
\value{
aggregated weekly data of the generic confirmed cases data (filtered by date if any)
}
\description{
\code{get_weekly_aggregated_data()} performs data transformation in the following steps:
\enumerate{
\item Group the weekly data by date (week date).
\item Aggregate the number of confirmed cases by week date.
\item Select only the week date and confirmed cases column.
\item Filter the data by given start and end date
}

The input dataframe \code{generic_data} must have the following columns:
\itemize{
\item \verb{<date name>}: date column (e.g. as.Date('2022-01-01')).
\item \verb{<cases count name>}: Confirmed Cases Count (e.g. 1, 2, ...).
}

Note that these columns can be defined in a generic name, and inputted as
the other two function parameters for data transformation (\code{date_column},
\code{number_column})

Assume the date column is the start of the epiweek.
}
\examples{
weekly_aggregated_data <- get_weekly_aggregated_data(
  generic_data,
  "date_of_report", "flu_a", "2022-10-16", "2023-12-31"
)
}
