% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation_generic.R
\name{get_weekly_data}
\alias{get_weekly_data}
\title{Extract Weekly Data from a Generic Dataframe}
\usage{
get_weekly_data(generic_data, date_column, number_column)
}
\arguments{
\item{generic_data}{data with a self defined column name for both date and cases count}

\item{date_column}{date column name str}

\item{number_column}{cases count column name str}
}
\value{
weekly data of the generic confirmed cases data
}
\description{
\code{get_weekly_data()} performs data transformation in the following steps:
\enumerate{
\item Create the week date column using \code{floor_date()}.
\item Select only the week date and confirmed cases column.
}

The input dataframe \code{generic_data} must have the following columns:
\itemize{
\item \verb{<date name>}: date column (e.g. as.Date('2022-01-01'))
\item \verb{<cases count name>}: Confirmed Cases Count (e.g. 1, 2, ...)
Note that these columns can be defined in a generic name, and inputted as
the other two function parameters for data transformation (\code{date_column},
\code{number_column})
}
}
\examples{

generic_data <- data.frame(
  date_of_report = as.Date(c(
    "2022-10-10", "2022-10-10",
    "2022-10-17", "2022-10-23", "2022-10-30",
    "2022-11-06", "2022-11-13", "2022-11-20",
    "2022-11-27", "2022-11-28"
  )),
  flu_a = c(7, 10, 32, 38, 43, 45, 73, 88, 44, 51),
  flu_b = c(6, 13, 31, 45, 50, 52, 68, 83, 45, 49)
)

weekly_data <- get_weekly_data(generic_data, "date_of_report", "flu_a")
}
