% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation_phrdw.R
\name{get_weekly_phrdw}
\alias{get_weekly_phrdw}
\title{get_weekly_phrdw - obtain weekly phrdw data group by date & age}
\usage{
get_weekly_phrdw(phrdw_flu_daily_count)
}
\arguments{
\item{phrdw_flu_daily_count}{raw phrdw data}
}
\value{
aggregated phrdw data by date & age
}
\description{
get_weekly_phrdw - obtain weekly phrdw data group by date & age
}
\examples{
get_weekly_phrdw(data.frame(
lis_date_collection = as.Date(c('2023-01-08','2023-01-08','2023-01-08','2023-01-09')),
                             result_lab_name = c("lab_a", "lab_b", "lab_a", "lab_b"),
                             age_years = c(0, 1, 2, 3),
                             sars_cov2 = c(1,2,3,4),
                             rsv = c(1,2,3,4),
                             flu_a = c(1,2,3,4),
                             flu_b = c(3,4,5,6)))
}
