% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation_phrdw.R
\name{get_weekly_phrdw_by_type_date_age}
\alias{get_weekly_phrdw_by_type_date_age}
\title{get_weekly_phrdw_by_type_date_age - obtain weekly phrdw data filtered by disease type, date, age}
\usage{
get_weekly_phrdw_by_type_date_age(
  weekly_phrdw_data,
  type,
  start_date,
  end_date,
  start_age = 0,
  end_age = 150
)
}
\arguments{
\item{weekly_phrdw_data}{weekly phrdw data from get_weekly_phrdw(phrdw_flu_daily_count)}

\item{type}{disease type (e.g. 'sars_cov2', 'rsv', 'flu_a', 'flu_b')}

\item{start_date}{start date string (e.g. '2022-01-01')}

\item{end_date}{end date string (e.g. '2022-12-31')}

\item{start_age}{start age integer (default = 0)}

\item{end_age}{end age integer (default = 150)}
}
\value{
weekly phrdw data filtered by disease type, date, age
}
\description{
get_weekly_phrdw_by_type_date_age - obtain weekly phrdw data filtered by disease type, date, age
}
\examples{
get_weekly_phrdw_by_type_date_age(
weekly_phrdw_data = get_weekly_phrdw(data.frame(
lis_date_collection = as.Date(c('2023-01-08','2023-01-08','2023-01-08','2023-01-09')),
                                 result_lab_name = c("lab_a", "lab_b", "lab_a", "lab_b"),
                                 age_years = c(0, 1, 2, 3),
                                 sars_cov2 = c(1,2,3,4),
                                 rsv = c(1,2,3,4),
                                 flu_a = c(1,2,3,4),
                                 flu_b = c(3,4,5,6))),
                             type='rsv',
                             start_date='2023-01-01',
                             end_date='2023-02-01',
                             start_age=0,
                             end_age=10)
}
