% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation_plover.R
\name{get_weekly_plover}
\alias{get_weekly_plover}
\title{Obtain the weekly cases from plover data}
\usage{
get_weekly_plover(plover_data)
}
\arguments{
\item{plover_data}{raw plover data before any transformation}
}
\value{
plover data aggregated by epiweek date, year, week, flu type.
}
\description{
\code{get_weekly_plover()} performs data transformation in the following steps:
\enumerate{
\item Uses \code{pivot_longer()} to convert the \code{flu_a} & \code{flu_b} columns into \code{flu_cases}
and its name into \code{type}.
\item Aggregates the sum of \code{flu_cases} by \code{epiWeek_date}, \code{epiWeek_year}, \code{Epiweek}, \code{type}.
\item Uses \code{pivot_wider()} to convert the \code{flu_cases} into the \code{flu_a} & \code{flu_b} columns.
}

The input dataframe \code{plover_data} must have the following columns:
\itemize{
\item \code{epiWeek_date}: Epiweek Date String (e.g. '2019-01-01')
\item \code{epiWeek_year}: Epiweek Year Number (e.g. 2019)
\item \code{Epiweek}: Epiweek Number (e.g. 1, 2, ..., 53)
\item \code{flu_a}: Confirmed Cases Count (e.g. 1, 2, ...)
\item \code{flu_b}: Confirmed Cases Count (e.g. 1, 2, ...)
}
}
\examples{
plover_data <- data.frame(
          epiWeek_date = as.Date(c('2022-10-02', '2022-10-09',
                     '2022-10-16', '2022-10-23', '2022-10-30',
                     '2022-11-06', '2022-11-13', '2022-11-20',
                      '2022-11-27', '2022-12-04')),
          epiWeek_year = c(2022,2022,2022,2022,
                      2022,2022,2022,2022,2022,2022),
           Epiweek = c(40,41,42,43,44,45,46,47,48,49),
           flu_a = c(17,19,32,38,43,45,73,88,94,105),
           flu_b = c(24,31,39,45,50,52,68,83,89,97))

get_weekly_plover(plover_data)
}
