% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation_plover.R
\name{get_weekly_plover_by_date_type}
\alias{get_weekly_plover_by_date_type}
\title{Filter the weekly plover data by date & disease type}
\usage{
get_weekly_plover_by_date_type(plover_data, type, start_date, end_date)
}
\arguments{
\item{plover_data}{raw plover data before any transformation}

\item{type}{disease type string (e.g. 'flu_a', 'flu_b')}

\item{start_date}{start date string (e.g. '2022-01-01')}

\item{end_date}{end date string (e.g. '2022-12-31')}
}
\value{
filtered weekly_plover_data
}
\description{
\code{get_weekly_plover_by_date_type()} filter the weekly plover data by a given
disease type and the date range, it performs the following steps:
\enumerate{
\item Filters the weekly plover data by the given epiweek date range.
\item Selects the epiweek date & flu type columns.
\item Renames the epiweek date as \code{date} & the flu type as \code{confirm}.
}

The input flu type must be 'flu_a' or 'flu_b'.

The input start date must be earlier than the input end date.

The input dataframe \code{plover_data} must have the following columns:
\itemize{
\item \code{epiWeek_date}: Epiweek Date String (e.g. '2019-01-01')
\item \code{epiWeek_year}: Epiweek Year Number (e.g. 2019)
\item \code{Epiweek}: Epiweek Number (e.g. 1, 2, ..., 53)
\item \code{flu_a}: Confirmed Cases Count (e.g. 1, 2, ...)
\item \code{flu_b}: Confirmed Cases Count (e.g. 1, 2, ...)
}
}
\examples{
get_weekly_plover_by_date_type(
  plover_data,
  "flu_a",
  "2022-10-01",
  "2022-12-05"
)
}
