% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phrdw_data}
\alias{phrdw_data}
\title{Daily/Weekly flu, RSV, sars_cov2 Cases from Phrdw}
\format{
A data frame with over 8000 rows and 7 variables:
\describe{
\item{lis_date_collection}{the date of that epiweek (2021-08-31 -- 2023-10-17)}
\item{result_lab_name}{the year of that epiweek (e.g. "BC Centre for Disease Control")}
\item{age_years}{the epiweek number (0 -- 103)}
\item{sars_cov2}{number of confirmed sars_cov2 cases}
\item{rsv}{number of confirmed RSV cases}
\item{flu_a}{number of confirmed flu A cases}
\item{flu_b}{number of confirmed flu B cases}
}
}
\usage{
phrdw_data
}
\description{
A dataset containing the flu, RSV, sars_cov2 Cases for different
labs & ages in BC Canada.
The variables are as follows:
}
\keyword{datasets}
