% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_functions.R
\name{plot.forecast_time_period}
\alias{plot.forecast_time_period}
\title{Plot forecasts at each iteration with uncertainty quantile ranges}
\usage{
\method{plot}{forecast_time_period}(x, time_period = NULL, ...)
}
\arguments{
\item{x}{object of class \code{forecast_time_period}}

\item{time_period}{optional parameter to show only plot at a specific time-point}

\item{...}{pass optional parameters to plot method}
}
\value{
Multiple plots with forecasts at each sliding window
}
\description{
Plot forecasts at each iteration with uncertainty quantile ranges
}
\examples{
plot(daily_time_period_result)
}
