% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_epiestim_model.R
\name{plot_all_time_period_forecast_data}
\alias{plot_all_time_period_forecast_data}
\title{Plot Individual Projection Data from A List of Time Period Result}
\usage{
plot_all_time_period_forecast_data(time_period_result)
}
\arguments{
\item{time_period_result}{output from experiment_time_period_epiestim()}
}
\value{
multiple plots
}
\description{
Plot Individual Projection Data from A List of Time Period Result
}
