% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_R_fit_comparison}
\alias{plot_R_fit_comparison}
\title{Compare Reproduction Number Estimates from Multiple EpiEstim Outputs}
\usage{
plot_R_fit_comparison(...)
}
\arguments{
\item{...}{Named EpiEstim output objects. Each object should contain a data frame
\code{R}}
}
\value{
A ggplot object showing median \(R\) estimates with 90\% credible intervals
(as ribbons), colored by each result.
}
\description{
This function creates a ggplot comparing the estimated reproduction number \(R\)
over time from multiple EpiEstim results. Each input should be an object from
EpiEstim containing an \code{R} dataframe.
}
\examples{
\dontrun{
# Assuming `fit1` and `fit2` are valid EpiEstim outputs with R data frames:
plot_R_fit_comparison(Fit1 = fit1, Fit2 = fit2)
}

}
