% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_rt}
\alias{plot_rt}
\title{Plot Mean Rt with time index (dates)}
\usage{
plot_rt(time_period_result)
}
\arguments{
\item{time_period_result}{output from  \code{forecast_time_period}}
}
\value{
Mean Rt with time index plot
}
\description{
Plot Mean Rt with time index (dates)
}
