% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_epiestim_model.R
\name{plot_violin_time_period_results_epiestim}
\alias{plot_violin_time_period_results_epiestim}
\title{Plot Violin Plot (EpiEstim), with option to plot specific pred_horizon}
\usage{
plot_violin_time_period_results_epiestim(
  time_period_result,
  pred_horizon_str = "",
  filter_draw = TRUE
)
}
\arguments{
\item{time_period_result}{output from experiment_time_period_epiestim()}

\item{pred_horizon_str}{str value of a pred_horizon (e.g. "1 Week Ahead")}

\item{filter_draw}{boolean value for extracting the draw by p25-p75 (if TRUE)}
}
\description{
Plot Violin Plot (EpiEstim), with option to plot specific pred_horizon
}
