% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pred_interval_forecast}
\alias{pred_interval_forecast}
\title{Calculate 95\% prediction interval bounds for smoothed predictions}
\usage{
pred_interval_forecast(time_period_result, data)
}
\arguments{
\item{time_period_result}{output from  \code{forecast_time_period}}

\item{data}{original input \emph{data frame} at each time-step containing two columns: date and confirm (number of cases per day)}
}
\value{
upper and lower bounds for 95\% prediction interval for smoothed predictions
}
\description{
Calculate 95\% prediction interval bounds for smoothed predictions
}
