% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_val_functions.R
\name{pred_samples_with_quantile_helper}
\alias{pred_samples_with_quantile_helper}
\title{Helper function to extract dataframe with week ahead forecasts for Violin plot}
\usage{
pred_samples_with_quantile_helper(tp, aggregate_unit = NULL)
}
\arguments{
\item{tp}{element from list output generated by \code{forecast_time_period_epiestim}}

\item{aggregate_unit}{Time forecasted samples are aggregated by (weekly or daily)}
}
\value{
Dataframe of forecasts at a single time-point (single element from list input)
}
\description{
Helper function to extract dataframe with week ahead forecasts for Violin plot
}
