% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rtrunc_norm}
\alias{rtrunc_norm}
\title{Sample from a truncated normal using inverse transform uniform sampling}
\usage{
rtrunc_norm(n, mean, sd, lower_lim = 0)
}
\arguments{
\item{n}{Number of random samples}

\item{mean}{Mean of distribution}

\item{sd}{Standard deviation of distribution}

\item{lower_lim}{Lower limit for truncation}
}
\description{
Sample from a truncated normal using inverse transform uniform sampling
}
