% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_functions.R
\name{summary.forecast_time_period}
\alias{summary.forecast_time_period}
\title{Generate a summary of model validation}
\usage{
\method{summary}{forecast_time_period}(object, pred_horizon_str = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{forecast_time_period}}

\item{pred_horizon_str}{A \emph{string} indicating the prediction horizon time period to plot}

\item{...}{Additional optional parameters to pass to the summary method}
}
\value{
A list containing two dataframes: \code{individual_quantiles} and \code{quantiles_summary} respectively:
\describe{
\item{date}{Daily or weekly date over which samples were aggregated}
\item{data_quantile}{Prediction quantile in which the confirmed case from data is located}
}
\describe{
\item{data_quantile}{Data quantiles}
\item{count}{Number of data points in one of the prediction quantiles (Q1, Q2, Q3, Q4)}
}
}
\description{
Outputs a summary indicating the prediction quantile in which the confirmed weekly case from the data falls.
The function warns when a confirmed case data-point is not within the predicted 95\% confidence interval
}
\examples{
summary(weekly_time_period_result, pred_horizon_str = "1 week ahead")
}
