% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_functions.R
\name{summary.forecast_time_period_epiestim}
\alias{summary.forecast_time_period_epiestim}
\title{Generate a summary of model validation}
\usage{
\method{summary}{forecast_time_period_epiestim}(object, pred_horizon_str = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{forecast_time_period_epiestim}}

\item{pred_horizon_str}{A \emph{string} indicating the prediction horizon time period to plot}

\item{...}{Additional optional parameters to pass to the summary method}
}
\value{
A list containing two dataframes: \code{individual_quantiles} and \code{quantiles_summary} respectively:
\describe{
\item{date}{Daily or weekly date over which samples were aggregated}
\item{data_quantile}{Prediction quantile in which the confirmed case from data is located}
}
\describe{
\item{data_quantile}{Data quantiles}
\item{count}{Number of data points in one of the prediction quantiles (Q1, Q2, Q3, Q4)}
}
}
\description{
Outputs a summary indicating the prediction quantile in which the confirmed weekly case from the data falls.
The function warns when a confirmed case data-point is not within the predicted quantiles.
}
\examples{
plover_data <- data.frame(
  epiWeek_date = as.Date(c(
    "2022-10-02", "2022-10-09",
    "2022-10-16", "2022-10-23", "2022-10-30",
    "2022-11-06", "2022-11-13", "2022-11-20",
    "2022-11-27", "2022-12-04"
  )),
  epiWeek_year = c(
    2022, 2022, 2022, 2022,
    2022, 2022, 2022, 2022, 2022, 2022
  ),
  Epiweek = c(40, 41, 42, 43, 44, 45, 46, 47, 48, 49),
  flu_a = c(17, 19, 32, 38, 43, 45, 73, 88, 94, 105),
  flu_b = c(24, 31, 39, 45, 50, 52, 68, 83, 89, 97)
)

weekly_plover_data <- get_weekly_plover(plover_data)

plover_dat_clean <- get_weekly_plover_by_date_type(
  weekly_plover_data,
  "flu_a",
  "2022-10-01",
  "2022-12-05"
)

time_period_result <- forecast_time_period_epiestim(
  data = plover_dat_clean,
  start_date_str = "2022-10-02", n_days = 14, type = "flu_a", aggregate_week = TRUE,
)
summary(time_period_result, pred_horizon_str = "1 week ahead")
}
