% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary_ind_quantiles_formatter}
\alias{summary_ind_quantiles_formatter}
\title{Format summary table for vriforecasting report}
\usage{
summary_ind_quantiles_formatter(summary_individual_quantiles)
}
\arguments{
\item{summary_individual_quantiles}{Individual quantile summary extracted from \code{summary} function}
}
\value{
Formatted summary table with wide format for coverage
}
\description{
Format summary table for vriforecasting report
}
