% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_weighted_diff}
\alias{time_weighted_diff}
\title{Extract squared error between predicted and confirmed case values weighted by number of time-points used to make prediction}
\usage{
time_weighted_diff(confirm, p50, pred_horizon_str = NULL)
}
\arguments{
\item{confirm}{confirmed weekly cases}

\item{p50}{prediction median quantile value}

\item{pred_horizon_str}{\emph{string} prediction horizon time period to plot}
}
\value{
\emph{numeric} Weighted squared error at each data time-point
}
\description{
Extract squared error between predicted and confirmed case values weighted by number of time-points used to make prediction
}
