% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validation_summary_text}
\alias{validation_summary_text}
\title{Print out text output for ViroReportR report detailing validation information including proportion of predictions in 50\% and 95\% intervals}
\usage{
validation_summary_text(time_period_result)
}
\arguments{
\item{time_period_result}{output from  \code{forecast_time_period}}
}
\value{
current forecast metrics
}
\description{
Print out text output for ViroReportR report detailing validation information including proportion of predictions in 50\% and 95\% intervals
}
