% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validation_summary_text_season}
\alias{validation_summary_text_season}
\title{Print out text output for ViroReportR report detailing validation information including proportion of predictions in 50\% and 95\% intervals}
\usage{
validation_summary_text_season(time_period_result, year_input = 2021)
}
\arguments{
\item{time_period_result}{output from  \code{forecast_time_period}}

\item{year_input}{\emph{numeric} indicating the year to generate the validation summary}
}
\value{
current forecast metrics
}
\description{
Print out text output for ViroReportR report detailing validation information including proportion of predictions in 50\% and 95\% intervals
}
