% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViroReportR-package.R
\docType{package}
\name{ViroReportR-package}
\alias{ViroReportR}
\alias{ViroReportR-package}
\title{ViroReportR: Miscellenous tools for generating forecasts from VRI data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

BCCDC R package for forecasting viral respiratory infection cases and outcomes.
}
\author{
\strong{Maintainer}: Mike Irvine \email{mike.irvine@bccdc.ca}

Authors:
\itemize{
  \item Caesar Wong
  \item Nirupama Tamvada
  \item Rebeca Falcao
  \item Nelson Tang
}

}
\keyword{internal}
