% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vriforecasting-package.R
\docType{package}
\name{vriforecasting-package}
\alias{vriforecasting}
\alias{vriforecasting-package}
\title{vriforecasting: Miscellenous tools for generating forecasts from VRI data}
\description{
BCCDC R package for forecasting viral respiratory infection cases and outcomes.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
