% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weekly_data}
\alias{weekly_data}
\title{Weekly respirartory viral cases}
\format{
A data frame with over 1000 rows and 5 variables:
\describe{
\item{epiWeek_date}{the date of that epiweek (2019-09-08 -- 2023-10-15)}
\item{epiWeek_year}{the year of that epiweek (2019 -- 2023)}
\item{Epiweek}{the epiweek number (1 -- 53)}
\item{flu_a}{number of confirmed flu A cases}
\item{flu_b}{number of confirmed flu B cases}
\item{corona}{number of confirmed SARS CoV2 cases}
\item{rsv}{number of confirmed RSV cases}
}
}
\usage{
weekly_data
}
\description{
A dataset containing weekly cases of Flu-A, Flu-B, SARS-CoV2 and RSV
The variables are as follows:
}
\keyword{datasets}
