% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weekly_transformed_plover_data}
\alias{weekly_transformed_plover_data}
\title{Weekly aggregated and transformed Flu Cases from Plover}
\format{
Data frame with
\describe{
\item{date}{Weekly date over which samples were aggregated}
\item{confirm}{confirmed weekly cases}
}
}
\source{
data produced by script \code{weekly_transformed_plover_data.R}
}
\usage{
weekly_transformed_plover_data
}
\description{
A dataset containing the flu cases in BC Canada aggregated by week after running package PLOVER transformation functions
The variables are as follows:
}
\keyword{datasets}
