% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{common_reliable_estimation_date}
\alias{common_reliable_estimation_date}
\title{Extract common date that all diseases in PHRDW and PLOVER data can be reliably estimated with in EpiEstim}
\usage{
common_reliable_estimation_date(
  PLOVER_data,
  PHRDW_data,
  flua_min = config$min_nb_cases_flua,
  flub_min = config$min_nb_cases_flub,
  cov_min = config$min_nb_cases_covid,
  rsv_min = config$min_nb_cases_rsv
)
}
\arguments{
\item{confirm}{confirmed weekly cases}

\item{p50}{prediction median quantile value}

\item{pred_horizon_str}{\emph{string} prediction horizon time period to plot}
}
\value{
\emph{numeric} Weighted squared error at each data time-point
}
\description{
Extract common date that all diseases in PHRDW and PLOVER data can be reliably estimated with in EpiEstim
}
