% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_val_functions.R
\name{create_forecast_df}
\alias{create_forecast_df}
\title{Extract data-frame with forecasts for validation violin plot}
\usage{
create_forecast_df(time_period_result)
}
\arguments{
\item{time_period_result}{output from  \code{forecast_time_period_epiestim}}
}
\value{
Data frame with forecasts at all time-points
}
\description{
Extract data-frame with forecasts for validation violin plot
}
\examples{
plover_data <- data.frame(
          epiWeek_date = as.Date(c('2022-10-02', '2022-10-09',
                     '2022-10-16', '2022-10-23', '2022-10-30',
                     '2022-11-06', '2022-11-13', '2022-11-20',
                      '2022-11-27', '2022-12-04')),
          epiWeek_year = c(2022,2022,2022,2022,
                      2022,2022,2022,2022,2022,2022),
           Epiweek = c(40,41,42,43,44,45,46,47,48,49),
           flu_a = c(17,19,32,38,43,45,73,88,94,105),
           flu_b = c(24,31,39,45,50,52,68,83,89,97))

weekly_plover_data <- get_weekly_plover(plover_data)

plover_dat_clean <- get_weekly_plover_by_date_type(
                             weekly_plover_data,
                             'flu_a',
                             '2022-10-01',
                             '2022-12-05')

time_period_result <- forecast_time_period_epiestim(data = plover_dat_clean,
start_date_str = "2022-10-02", n_days = 14, type = "flu_a")

create_forecast_df(time_period_result)


}
