% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_forecast_df}
\alias{create_forecast_df}
\title{Extract data-frame with forecasts for validation violin plot}
\usage{
create_forecast_df(time_period_result)
}
\arguments{
\item{time_period_result}{output from  \code{forecast_time_period}}
}
\value{
Data frame with forecasts at all time-points
#' \describe{
\item{date}{daily date or weekly date over which samples were aggregated}
\item{p50}{median quantile value}
\item{p25}{quantile value of probability 0.25}
\item{p75}{quantile value of probability 0.75}
\item{p025}{quantile value of probability 0.05}
\item{p975}{quantile value of probability 0.95}
\item{pred_horizon}{Prediction time horizon (time period ahead for which prediction was made)}
\item{value}{prediction values from simulated draws}
}
}
\description{
Extract data-frame with forecasts for validation violin plot
}
