% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_quantiles}
\alias{create_quantiles}
\title{summarise a data frame \code{d} by groups along a \code{variable}}
\usage{
create_quantiles(d, ..., variable = NULL)
}
\arguments{
\item{d}{tibble data frame}

\item{...}{group_by variables}

\item{variable}{string}
}
\value{
Data frame containing sample quantiles at probabilities 0.05, 0.25, 0.50,
0.75 and 0.95
}
\description{
summarise a data frame \code{d} by groups along a \code{variable}
}
