% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{current_forecast_text}
\alias{current_forecast_text}
\title{Print out text output for vriforecasting report detailing current number of case visits, last value of Rt and corresponding intervals}
\usage{
current_forecast_text(time_period_result, ...)
}
\arguments{
\item{time_period_result}{output from  \code{forecast_time_period}}

\item{...}{optional arguments to be passed on to \code{forecast_metrics}}
}
\value{
current forecast metrics
}
\description{
Print out text output for vriforecasting report detailing current number of case visits, last value of Rt and corresponding intervals
}
