% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extend_rows_model_data}
\alias{extend_rows_model_data}
\title{extract model data with extension during each iteration of loop}
\usage{
extend_rows_model_data(data, min_model_date_str, extension_interval = 1)
}
\arguments{
\item{data}{\emph{data frame} containing two columns: date and confirm (number of cases per day)}

\item{min_model_date_str}{start date (in str)}

\item{extension_interval}{an integer (# of days)}
}
\value{
input model_data subset to extension interval period specified
}
\description{
extract model data with extension during each iteration of loop
}
