% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_epiestim_model.R
\name{fit_epiestim_model}
\alias{fit_epiestim_model}
\title{fit_epiestim_model - Function to estimate the reproduction number of an epidemic from weekly data}
\usage{
fit_epiestim_model(
  data,
  dt = 7L,
  type = NULL,
  mean_si = NULL,
  std_si = NULL,
  recon_opt = "match",
  method = "parametric_si",
  ...
)
}
\arguments{
\item{data}{\emph{data frame} containing two columns: date and confirm (number of cases per day)}

\item{dt}{\emph{Integer} Number of days aggregated (set to 7 by default for weekly aggregate)}

\item{type}{\emph{character} Specifies type of epidemic. Must be one of "Influenza", "RSV" and "COVID"}

\item{mean_si}{\emph{Numeric} User specification of mean of parametric serial interval}

\item{std_si}{\emph{Numeric} User specification of standard deviation of parametric serial interval}

\item{...}{Optional parameters to pass on to {\code{\link[EpiEstim]{estimate_R}}} (see help page) to control estimation of reproduction number}
}
\value{
Object of class {\code{\link[EpiEstim]{estimate_R}}} (see \code{EpiEstim} help page)
}
\description{
A wrapper function for {\code{\link[EpiEstim]{estimate_R}}} from the \code{EpiEstim} library to estimate the reproduction number of epidemics to support short-term forecasts
}
\details{
\code{fit_epiestim_model} currently supports the following epidemics: Influenza, RSV and COVID-19. The serial intervals for the estimation of R were retrieved from
\link{Cowling et al., 2011}, \link{Vink et al., 2014} and \link{Madewell et al., 2023} for Influenza, RSV and COVID (BA.5 Omicron variant) respectively
}
