% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_epiestim_model_functions.R
\name{forecast_time_period_epiestim}
\alias{forecast_time_period_epiestim}
\title{Iterate through a time-period as a sliding window to produce short-term forecasts with the EpiEstim model fit}
\usage{
forecast_time_period_epiestim(
  data,
  start_date,
  n_days = 7,
  time_period = "daily",
  type = NULL,
  verbose = FALSE,
  smoothing_cutoff = 10,
  ...
)
}
\arguments{
\item{data}{\emph{data frame} containing two columns: date and confirm (number of cases per week)}

\item{start_date}{Initial starting time-point. Must match a timepoint in the input dataset}

\item{n_days}{Number of days to forecast ahead. Defaults to 7}

\item{time_period}{time period string (e.g. 'daily', 'weekly'). Default is daily}

\item{type}{\emph{character} Specifies type of epidemic. Must be one of "flu_a", "flu_b", "rsv", "sars_cov2" or "other"}

\item{verbose}{set to true to display progress output}

\item{smoothing_cutoff}{number of time periods windows after to start smoothing}

\item{...}{Pass on optional arguments from \code{fit_epiestim_model}}
}
\value{
List of class \code{forecast_time_period}
storing quantiles of both daily and weekly forecasts from each sliding window
}
\description{
Function to produce short-term forecasts from objects of class {\code{\link[EpiEstim]{estimate_R}}}
}
\examples{

#  Daily forecast
forecast_time_period_epiestim(
  data = weekly_transformed_plover_data,
  start_date = "2022-10-02", n_days = 14, type = "flu_a"
)

forecast_time_period_epiestim(
  data = weekly_transformed_plover_data,
  start_date = "2022-10-02", n_days = 14, type = "flu_a", time_period = "weekly"
)
}
