% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_epiestim_model_functions.R
\name{forecast_time_period_epiestim}
\alias{forecast_time_period_epiestim}
\title{Iterate through a time-period as a sliding window to produce short-term forecasts with the EpiEstim model fit}
\usage{
forecast_time_period_epiestim(
  data,
  start_date_str,
  n_days = 7,
  aggregate_week = FALSE,
  type = NULL,
  ...
)
}
\arguments{
\item{data}{\emph{data frame} containing two columns: date and confirm (number of cases per day)}

\item{start_date_str}{Initial starting time-point. Must match a timepoint in the input dataset}

\item{n_days}{The number of days to run simulations for. Defaults to 7}

\item{aggregate_week}{\emph{logical} argument specifying whether to aggregate forecasts by weekly quantiles. Default is to return daily quantiles}

\item{type}{\emph{character} Specifies type of epidemic. Must be one of "Influenza", "RSV" and "COVID"}

\item{...}{Pass on optional arguments from \code{fit_epiestim_model}}
}
\value{
List storing quantiles of both daily and 2 week ahead weekly forecasts from each sliding window
}
\description{
Function to produce short-term forecasts from objects of class {\code{\link[EpiEstim]{estimate_R}}}
}
