% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_forecast_report.R
\name{generate_forecast_report}
\alias{generate_forecast_report}
\title{Generate forecast report - Function to generate and knit report of forecasts for all supported viral respiratory diseases (Flu-A, Flu-B, RSV and SARS-CoV2)}
\usage{
generate_forecast_report(output_dir = NULL)
}
\arguments{
\item{output_dir}{specify path to output directory that HTML report should be rendered in}
}
\description{
Function to render a full season viral respiratory report as an HTML document in the directory of your choice. The report includes plots of the observed trends and
model produced forecasts
}
