% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation_phrdw.R
\name{get_daily_phrdw}
\alias{get_daily_phrdw}
\title{Obtain daily phrdw data group by date & age}
\usage{
get_daily_phrdw(phrdw_flu_daily_count)
}
\arguments{
\item{phrdw_flu_daily_count}{raw phrdw data}
}
\value{
An aggregated daily phrdw data by date & age
}
\description{
\code{get_daily_phrdw()} performs data transformation in the following steps:
\enumerate{
\item Uses \code{filter()} to remove those missing \code{result_lab_name}.
\item set the \code{age_years} as numeric.
\item Aggregates the confirmed cases for each diseases by \code{lis_date_collection}, \code{age_years}.
}

The input dataframe \code{phrdw_flu_daily_count} must have the following columns:
\itemize{
\item \code{lis_date_collection}: Date of collection (e.g. '2019-01-01')
\item \code{result_lab_name}: The name of the result lab (e.g. 'BCCDC')
\item \code{age_years}: Age (e.g. 1, 2, ...)
\item \code{sars_cov2}: Confirmed Cases Count (e.g. 1, 2, ...)
\item \code{rsv}: Confirmed Cases Count (e.g. 1, 2, ...)
\item \code{flu_a}: Confirmed Cases Count (e.g. 1, 2, ...)
\item \code{flu_b}: Confirmed Cases Count (e.g. 1, 2, ...)
}
}
\examples{
phrdw_data <- data.frame(
  lis_date_collection = as.Date(c(
    "2022-10-02", "2022-10-02",
    "2022-10-09", "2022-10-09",
    "2022-10-16", "2022-10-16",
    "2022-10-23", "2022-10-23",
    "2022-10-30", "2022-10-30"
  )),
  result_lab_name = c(
    "BC Centre for Disease Control", "BC Centre for Disease Control",
    "BC Centre for Disease Control", "BC Centre for Disease Control",
    "BC Centre for Disease Control", "BC Centre for Disease Control",
    "BC Centre for Disease Control", "BC Centre for Disease Control",
    "BC Centre for Disease Control", "BC Centre for Disease Control"
  ),
  age_years = c(7, 8, 7, 8, 7, 8, 7, 8, 7, 8),
  sars_cov2 = c(15, 17, 30, 28, 41, 43, 71, 74, 80, 88),
  rsv = c(14, 16, 29, 27, 40, 42, 70, 73, 79, 87),
  flu_a = c(13, 15, 28, 26, 39, 41, 69, 72, 78, 86),
  flu_b = c(12, 14, 27, 25, 38, 40, 68, 71, 77, 85)
)

get_daily_phrdw(phrdw_data)
}
