% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation_plover.R
\name{get_weekly_plover}
\alias{get_weekly_plover}
\title{get_weekly_plover - calculate the weekly cases from plover data}
\usage{
get_weekly_plover(plover_data)
}
\arguments{
\item{plover_data}{raw plover data before any transformation}
}
\value{
weekly_plover_data plover data aggregated by epiweek date, year, week type
}
\description{
get_weekly_plover - calculate the weekly cases from plover data
}
\examples{
get_weekly_plover(data.frame(
epiWeek_date = as.Date(c('2019-09-08','2019-09-29','2019-10-06','2019-10-13')),
                             epiWeek_year = c(2019, 2019, 2019, 2019),
                             Epiweek = c(37,40,41,42),
                             flu_a = c(1,2,3,4),
                             flu_b = c(3,4,5,6)))
}
