% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation_plover.R
\name{get_weekly_plover_by_date_type}
\alias{get_weekly_plover_by_date_type}
\title{get_weekly_plover_by_date_type - filter the weekly plover data by date & disease type}
\usage{
get_weekly_plover_by_date_type(weekly_plover_data, type, start_date, end_date)
}
\arguments{
\item{weekly_plover_data}{weekly plover data from get_weekly_plover}

\item{type}{disease type string (e.g. "flu_a", "flu_b")}

\item{start_date}{start date string (e.g. '2022-01-01')}

\item{end_date}{end date string (e.g. '2022-12-31')}
}
\value{
filtered weekly_plover_data
}
\description{
get_weekly_plover_by_date_type - filter the weekly plover data by date & disease type
}
\examples{
plover_data <- data.frame(
epiWeek_date = as.Date(c('2019-09-08','2019-09-29','2019-10-06','2019-10-13')),
                             epiWeek_year = c(2019, 2019, 2019, 2019),
                             Epiweek = c(37,40,41,42),
                             flu_a = c(1,2,3,4),
                             flu_b = c(3,4,5,6))

weekly_plover_data <- get_weekly_plover(plover_data)

get_weekly_plover_by_date_type(
                             weekly_plover_data,
                             'flu_a',
                             '2019-09-01',
                             '2019-10-13')
}
