% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_rt}
\alias{plot_rt}
\title{Plot Mean Rt with time index (dates)}
\usage{
plot_rt(forecast_results)
}
\arguments{
\item{forecast_results}{is the output of \code{generate_forecast}.}
}
\value{
Mean Rt with time index plot
}
\description{
Plot Mean Rt with time index (dates)
}
\examples{
\dontrun{
# Assuming forecast_results is the output generated by 
#`generate_forecast()``, i.e. forecast for one disease.
plot_rt(forecast_results)
}
}
