% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_rt}
\alias{plot_rt}
\title{Plot Mean Rt with time index (dates)}
\usage{
plot_rt(forecast_results)
}
\arguments{
\item{forecast_results}{output from  \code{generate_forecast}}
}
\value{
Mean Rt with time index plot
}
\description{
Plot Mean Rt with time index (dates)
}
