% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pred_interval_forecast}
\alias{pred_interval_forecast}
\title{Calculate 95\% prediction interval bounds for smoothed predictions}
\usage{
pred_interval_forecast(time_period_result, pred_horizon_str = NULL)
}
\arguments{
\item{time_period_result}{output from  \code{forecast_time_period}}

\item{pred_horizon_str}{A \emph{string} indicating the prediction horizon time period to plot}
}
\value{
upper and lower bounds for 95\% prediction interval for smoothed predictions
}
\description{
Calculate 95\% prediction interval bounds for smoothed predictions
}
