% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_epiestim_model.R
\name{pred_samples_with_quantile_helper}
\alias{pred_samples_with_quantile_helper}
\title{Helper function to extract dataframe with 1 week and 2 week ahead forecasts for Violin plot}
\usage{
pred_samples_with_quantile_helper(tp)
}
\arguments{
\item{tp}{element from list output generated by \code{forecast_time_period_epiestim}}
}
\value{
Dataframe of 1 week and 2 week forecasts at a single timepoint
}
\description{
Helper function to extract dataframe with 1 week and 2 week ahead forecasts for Violin plot
}
