% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pred_samples_with_quantile_helper}
\alias{pred_samples_with_quantile_helper}
\title{Helper function to extract dataframe with week ahead forecasts for Violin plot}
\usage{
pred_samples_with_quantile_helper(tp, aggregate_unit = NULL)
}
\arguments{
\item{tp}{element from list output generated by \code{forecast_time_period_epiestim}}

\item{aggregate_unit}{Time forecasted samples are aggregated by (weekly or daily)}
}
\value{
Dataframe of forecasts at a single time-point (single element from list input)
\describe{
\item{date}{daily date or weekly date over which samples were aggregated}
\item{p50}{median quantile value}
\item{p25}{quantile value of probability 0.25}
\item{p75}{quantile value of probability 0.75}
\item{p05}{quantile value of probability 0.05}
\item{p95}{quantile value of probability 0.95}
\item{pred_horizon}{Prediction time horizon (time period ahead for which prediction was made)}
\item{value}{prediction values from simulated draws}
}
}
\description{
Helper function to extract dataframe with week ahead forecasts for Violin plot
}
