% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary_ind_quantiles_formatter}
\alias{summary_ind_quantiles_formatter}
\title{Format summary table for ViroReportR report}
\usage{
summary_ind_quantiles_formatter(time_period_result)
}
\arguments{
\item{time_period_result}{output from  \code{forecast_time_period}}
}
\value{
Formatted data.table summary table with wide format for coverage
}
\description{
Format summary table for ViroReportR report
}
