% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViroReportR-package.R
\docType{package}
\name{ViroReportR-package}
\alias{ViroReportR}
\alias{ViroReportR-package}
\title{ViroReportR: Respiratory Viral Infection Forecast Reporting}
\description{
Tools for reporting and forecasting viral respiratory infections, using case surveillance data. Report generation tools for short-term forecasts, and validation metrics for an arbitrary number of customizable respiratory viruses.
}
\author{
\strong{Maintainer}: Mike Irvine \email{mike.irvine@bccdc.ca} [copyright holder]

Authors:
\itemize{
  \item Caesar Wong
  \item Nirupama Tamvada
  \item Rebeca Falcao
  \item Nelson Tang
}

}
\keyword{internal}
