% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vriforecasting-package.R
\docType{package}
\name{vriforecasting-package}
\alias{vriforecasting}
\alias{vriforecasting-package}
\title{vriforecasting: Miscellenous tools for generating forecasts from VRI data}
\description{
BCCDC R package for forecasting viral respiratory infection cases and outcomes.
}
\author{
\strong{Maintainer}: Mike Irvine \email{mike.irvine@bccdc.ca}

Authors:
\itemize{
  \item Caesar Wong
  \item Nirupama Tamvada
}

}
\keyword{internal}
