% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weekly_time_period_result}
\alias{weekly_time_period_result}
\title{Time period result: weekly forecasts of PLOVER flu data in sliding windows as the output of \code{forecast_time_period_epiestim}}
\format{
List of class \code{forecast_time_period_epiestim}
storing quantiles of both daily and weekly forecasts from each sliding window
}
\source{
data produced by script \code{time_period_result.R}
}
\usage{
weekly_time_period_result
}
\description{
Time period result: weekly forecasts of PLOVER flu data in sliding windows as the output of \code{forecast_time_period_epiestim}
}
\keyword{datasets}
